python-server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
Werkzeug==3.0.1
pytesseract==0.3.10

